# Use FMA for cscal and zscal Haswell microkernels.