--- a/kernel/x86_64/cscal.c
+++ b/kernel/x86_64/cscal.c
@@ -337,6 +337,15 @@ int CNAME(BLASLONG n, BLASLONG dummy0, BLASLONG dummy1, FLOAT da_r, FLOAT da_i, 
 	}
 
 
+#ifdef HAVE_KERNEL
+	if ( da_r != 0.0 && da_i != 0.0 ) {
+		alpha[0] = da_r;
+		alpha[1] = da_i;
+		cscal_kernel(n , alpha, x);
+		return(0);
+	}
+#endif
+
 	BLASLONG n1 = n & -16;
 	if ( n1 > 0 )
 	{
@@ -352,8 +361,10 @@ int CNAME(BLASLONG n, BLASLONG dummy0, BLASLONG dummy1, FLOAT da_r, FLOAT da_i, 
 		else
 			if ( da_i == 0 )
 				cscal_kernel_16_zero_i(n1 , alpha , x);
+#ifndef HAVE_KERNEL
 			else
 				cscal_kernel_16(n1 , alpha , x);
+#endif
 
 		i = n1 << 1;
 		j = n1;