@bartoldeman 2022-10-22
I put this to draft since I bumped into https://github.com/xianyi/OpenBLAS/commit/22aa81f3e587c
The alternative is then to force no FMAs in the C code, of course, using e.g.
`#pragma GCC optimize ("fp-contract=off")`
from that patch or some other technique.