@martin-frbg 2022-10-22
oh right, I had a feeling of deja vue but was too tired to search yesterday. at this point I am no longer sure if we should even rely this much on the LAPACK testsuite or simply adjust its thresholds to allow "healthy" deviations.