@bartoldeman 2022-10-22
I played a bit with the thresholds but there were still some hard cases that fail there.
There is some annoying behaviour with using FMAs to multiply complex numbers, in particular as 22aa81f3e587c noted, when the imaginary part of the result is close to 0, e.g. with `(a+bi) * (a-bi) = a**2 + b**2`; without FMAs the imaginary part of the result is always 0, but with them you can end up with e.g. `-5.55112e-17` for `a=3.0` and `b=1/3.0` due to the asymmetry in `a*b-a*b` with FMAs (of course we know that a*b isn't exactly 1.0, that's well known numerically, but subtraction of two things rounded the same way still gives 0).

There are a few relevant places in LAPACK where they compute a * conj(a), and that doesn't help for sure, but for now it simply seems best to turn off FMAs for complex multiplications to avoid surprising behaviour.